requirements.txt
numpy==1.19.4
requests==2.24.0
autocorrect==2.2.2
colorthief==0.2.1
Pillow==8.0.1
pyautogui==0.9.52
pytesseract==0.3.6
